+++ provision/target/entrypoint.py
@@ -18,6 +18,7 @@ app = FastAPI(
     debug=False,
 )
 
+# CORS middleware
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["*"],