+++ CHANGELOG.md
@@ -9,3 +9,7 @@ Bump project version into `0.5.1`
 - [x] (application) Build application logic within development
 
 - [x] (docs) Updated the documentation for project (README, sad, changelog, etc.)
+
+- [x] Build application of probe (agent, collector) instance
+
+- [x] Build application for target (for HTTP call) with single `/_/health` endpoint